Initial commit
This pipeline is mostly for internal use. It starts with a folder of read files (fastq.gz) and returns USEARCH taxon tables and OTU tables. OTUs are BLASTed, and then we then use taxonkit to look up taxonomic information. You will need to manually add taxonomic info to the OTU table (pay attention to sort order).